Move nightshade classes to primitives
